api-autenticacao/requirements.txt
Flask
Flask-JWT-Extend
Flask-Mail
supabase
python-dotenv
werkzeug